% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_intermittent.R
\name{approximate_entropy}
\alias{approximate_entropy}
\title{Compute the entropy of the time series}
\usage{
approximate_entropy(x, m, r)
}
\arguments{
\item{x}{The time series.}

\item{m}{Length of compared run of data.}

\item{r}{Filtering level, must be positive.}
}
\value{
the value of the feature \code{Entropy}
}
\description{
The approximate entropy algorithm in package \code{TSEntropies} is used.
}
