% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors_intermittent.R
\name{calc_errors_mase}
\alias{calc_errors_mase}
\title{Calculate the MASE error}
\usage{
calc_errors_mase(dataset)
}
\arguments{
\item{dataset}{The dataset. A list including some time series.
\code{$x} is the historical value, \code{$xx} is the true value,
and \code{$ff} is the forecast values based on the 12 methods.}
}
\value{
\code{dataset} with MASE errors
}
\description{
Calculate the MASE error
}
