% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors_intermittent.R
\name{calc_errors_mase_nos}
\alias{calc_errors_mase_nos}
\title{Calculate a compositive error considering MASE and NOS errors}
\usage{
calc_errors_mase_nos(dataset, a = 0)
}
\arguments{
\item{dataset}{The dataset. A list including some time series.
In each time series, \code{x} is the historical value, \code{xx} is the true value,
and \code{ff} is the forecasting values based on the 12 methods.}

\item{a}{The penalty factor for NOS.}
}
\value{
\code{dataset} with MASE errors
}
\description{
Calculate a compositive error considering MASE and NOS errors
}
