% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors_intermittent.R
\name{calc_errors_rmsse}
\alias{calc_errors_rmsse}
\title{Calculate RMSSE errors of the given point forecasts}
\usage{
calc_errors_rmsse(dataset)
}
\arguments{
\item{dataset}{The dataset. A list including some time series.
In each time series, \code{x} is the historical value, \code{xx} is the true value,
and \code{ff} is point forecasts based on a series forecasting methods.}
}
\value{
\code{dataset} with RMSSE errors
}
\description{
Calculate RMSSE errors of the given point forecasts
}
