% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors_intermittent.R
\name{calc_errors_spl}
\alias{calc_errors_spl}
\title{Calculate SPL errors of the given quantile forecasts}
\usage{
calc_errors_spl(dataset, u = 0.995)
}
\arguments{
\item{dataset}{The dataset. A list including some time series.
In each time series, \code{x} is the historical value, \code{xx} is the true value,
and \code{quantile_forec} is the quantile forecasts based on a series forecasting methods.}

\item{u}{The quantile.
Four high quantiles are considered, i.e. 0.750, 0.835, 0.975, and 0.995.}
}
\value{
\code{dataset} with SPL errors
}
\description{
Calculate SPL errors of the given quantile forecasts
}
