% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasts_intermittent.R
\name{calculate_forec_point}
\alias{calculate_forec_point}
\title{Function to calculate the forecasts of intermittent demand}
\usage{
calculate_forec_point(data, h = 12)
}
\arguments{
\item{data}{The dataset. A list including some time series.
\code{$x} is the historical value, and \code{$xx} is the true value.}

\item{h}{Forecasting horizon. 12 is the default.}
}
\value{
\code{data} with forecasts
}
\description{
Function to calculate the forecasts of intermittent demand
}
