% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_intermittent.R
\name{change_mean_abs}
\alias{change_mean_abs}
\title{#' Compute the feature \code{Change.quantile} of the time series
#'
#' The average, absolute value of consecutive changes of the series
#' inside a fixed range given by the quantiles.
#'
#' @param x The time series.
#' @param ql The lower quantile of the range.
#' @param qh The higher quantile of the range.
#' @param isabs Logical value. Should the absolute differences be taken?
#' @param f_agg The aggregator function (e.g. mean, var, std, median)
#'
#' @return the value of the feature \code{Change.quantile}
#' @export}
\usage{
change_mean_abs(x, isabs = T, f_agg = "mean")
}
\arguments{
\item{x}{The time series.}

\item{isabs}{Logical value. Should the absolute differences be taken?}

\item{f_agg}{The aggregator function (e.g. mean, var, std, median)}
}
\value{
the value of the feature \code{Change.mean.abs}
}
\description{
change_quantiles <- function(x, ql, qh, isabs, f_agg){
if (ql >= qh){
return(0)
}
div = diff(x)
if (isabs) {
div = abs(div)
}
bin_cat = quantile(x, probs = c(ql, qh))
bin_cat_0 = x>=bin_cat\link{1} & x<=bin_cat\link{2}
bin_cat_0_roll_1 = c(bin_cat_0\link{length(bin_cat_0)}, bin_cat_0\link{-length(bin_cat_0)})
ind = c(bin_cat_0 & bin_cat_0_roll_1)\link{-1}
aggregator = get(f_agg)
return(aggregator(div\link{ind}))
}
Compute the feature \code{Change.mean.abs} of the time series
}
\details{
The average, absolute value of consecutive changes of the series
}
