% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_intermittent.R
\name{change_quantiles}
\alias{change_quantiles}
\title{Compute the feature \code{Change.quantile} of the time series}
\usage{
change_quantiles(x, ql, qh, isabs, f_agg)
}
\arguments{
\item{x}{The time series.}

\item{ql}{The lower quantile of the range.}

\item{qh}{The higher quantile of the range.}

\item{isabs}{Logical value. Should the absolute differences be taken?}

\item{f_agg}{The aggregator function (e.g. mean, var, std, median)}
}
\value{
the value of the feature \code{Change.quantile}
}
\description{
The average, absolute value of consecutive changes of the series
inside a fixed range given by the quantiles.
}
