% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_intermittent.R
\name{compute_CV2}
\alias{compute_CV2}
\title{Computing the coefficient of variation squared (CV2) of non-zero demand}
\usage{
compute_CV2(x)
}
\arguments{
\item{x}{Time series.}
}
\value{
CV2.
}
\description{
Computing the coefficient of variation squared (CV2) of non-zero demand
}
