% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{compute_diversity}
\alias{compute_diversity}
\title{Function to compute the diversity}
\usage{
compute_diversity(data)
}
\arguments{
\item{data}{The dataset. A list including some time series, each of which has
a \code{ff} object with the forecasting values based on the 12 methods}
}
\value{
\code{data} with the diversity of forecasts (shown in \code{features}).
}
\description{
Function to compute the diversity
}
