% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_intermittent.R
\name{compute_IDI}
\alias{compute_IDI}
\title{Computing the averaged Inter-Demand Interval(IDI)}
\usage{
compute_IDI(x)
}
\arguments{
\item{x}{Time series.}
}
\value{
IDI.
}
\description{
Computing the averaged Inter-Demand Interval(IDI)
}
