% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_intermittent.R
\name{compute_ifeatures}
\alias{compute_ifeatures}
\title{Function to compute the nine features for intermittent demand}
\usage{
compute_ifeatures(
  ts,
  scale = TRUE,
  features = c("IDI", "CV2", "Entropy", "Percent.zero", "Percent.beyond.sigma",
    "Linear.chunk.var", "change_mean_abs", "Ratio.last.chunk", "Percent.zero.end"),
  chunk_len = 12,
  num_segments = 4
)
}
\arguments{
\item{ts}{Time series.}

\item{scale}{Logical value. If scale is TRUE, the ts is standardized.}

\item{features}{Names of selected features for intermittent demand.}
}
\value{
\code{features_out}
A matrix with one row and nine columns containing the nine features.
}
\description{
Function to compute the nine features for intermittent demand
}
