% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasts_intermittent.R
\name{CRO_forec_inter}
\alias{CRO_forec_inter}
\title{Forecast based on Croston??s method}
\usage{
CRO_forec_inter(
  ts,
  h = 12,
  quantile = F,
  u = c(0.005, 0.025, 0.165, 0.25, 0.5, 0.75, 0.835, 0.975, 0.995)
)
}
\arguments{
\item{ts}{The time series.}

\item{h}{Forecasting horizon. 12 is the default.}

\item{quantile}{Logical value. If quantile is TRUE, calculate the quantile forecasts.}

\item{u}{A vector containing the quantiles to calculate for forecasting.}
}
\value{
A \code{data.frame} with point forecasts and quantile forecasts.
}
\description{
Forecast based on Croston??s method
}
