% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasts_intermittent.R
\name{ensemble_forecast_quantile}
\alias{ensemble_forecast_quantile}
\title{Calculate the combined quantile forecast}
\usage{
ensemble_forecast_quantile(predictions, dataset, u = 0.995)
}
\arguments{
\item{predictions}{The combination weights for each time series,
obtained from \code{M4metalearning::predict_selection_ensemble}.}

\item{dataset}{The dataset. A list including some time series.
In each time series, \code{x} is the historical value, \code{xx} is the true value,
and \code{quantile_forec} is the quantile forecasts based on a series forecasting methods.}

\item{u}{The quantile.
Four high quantiles are considered, i.e. 0.750, 0.835, 0.975, and 0.995.}
}
\value{
\code{dataset} with combined quantile forecast.
}
\description{
Calculate the combined quantile forecast
}
