% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasts_intermittent.R
\name{forec_methods_inter}
\alias{forec_methods_inter}
\title{The methods used to calculate the forecasts of intermittent demand}
\usage{
forec_methods_inter()
}
\value{
A list including the names of forecasting methods.
}
\description{
The methods used to calculate the forecasts of intermittent demand
}
