% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_intermittent.R
\name{intervals}
\alias{intervals}
\title{Computing the intervals between the positive demands.}
\usage{
intervals(x)
}
\arguments{
\item{x}{Time series.}
}
\value{
A vector of intervals.
}
\description{
Computing the intervals between the positive demands.
}
