% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_intermittent.R
\name{linear_chunk_var}
\alias{linear_chunk_var}
\title{Compute the feature \code{Linear.chunk.var} of the time series}
\usage{
linear_chunk_var(x, f_agg, chunk_len = 10)
}
\arguments{
\item{x}{The time series.}

\item{f_agg}{The aggregator function (e.g. mean, var, std, median)}

\item{chunk_len}{The length of the component chunks in the time series.}
}
\value{
the value of the feature \code{Linear.chunk.var}
}
\description{
The coefficient of linear least-squares regression
for variances of component chunks in the time series.
}
