% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_intermittent.R
\name{percent_zero}
\alias{percent_zero}
\title{Computing the percentage of observations that are zero values}
\usage{
percent_zero(x, t = 0)
}
\arguments{
\item{x}{Time series.}
}
\value{
the value of the feature \code{Percent.zero}
}
\description{
Computing the percentage of observations that are zero values
}
