% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_intermittent.R
\name{ratio_last_chunk}
\alias{ratio_last_chunk}
\title{Compute the feature \code{Ratio.last.chunk} of the time series}
\usage{
ratio_last_chunk(x, num_segments)
}
\arguments{
\item{x}{The time series.}

\item{num_segments}{The number of chunks in the time series.}
}
\value{
the value of the feature \code{Ratio.last.chunk}
}
\description{
The ratio of the sum of squares of the last chunk to the whole series.
}
