% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors_intermittent.R
\name{summary_performance_error_intermittent}
\alias{summary_performance_error_intermittent}
\title{Calculate the forecasting errors of the proposed method}
\usage{
summary_performance_error_intermittent(
  dataset,
  error = c("sME", "sMAE", "sMSE", "MASE", "NOS", "sMPIS", "sMAPIS")
)
}
\arguments{
\item{dataset}{The dataset. A list including some time series.
\code{$x} is the historical value, \code{$xx} is the true value,
\code{$ff} is the forecasting values based on the 12 methods,
and \code{$y_hat} is the combined forecast of the proposed method.}

\item{error}{A vector including the errors to output.}
}
\value{
\code{dataset} with forecasting errors.
}
\description{
Calculate the forecasting errors of the proposed method
}
