% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors_intermittent.R
\name{summary_performance_rmsse}
\alias{summary_performance_rmsse}
\title{Calculate the RMSSE error of the proposed method}
\usage{
summary_performance_rmsse(dataset)
}
\arguments{
\item{dataset}{The dataset. A list including some time series.
In each time series, \code{x} is the historical value, \code{xx} is the true value,
and \code{y_hat} is the combined point forecast of the proposed method.}
}
\value{
The average RMSSE error of all time series.
}
\description{
Calculate the RMSSE error of the proposed method
}
