% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors_intermittent.R
\name{summary_performance_spl}
\alias{summary_performance_spl}
\title{Calculate the SPL error of the proposed method}
\usage{
summary_performance_spl(dataset, u = 0.995)
}
\arguments{
\item{dataset}{The dataset. A list including some time series.
In each time series, \code{x} is the historical value, \code{xx} is the true value,
and \code{y_hat} is the combined quantile forecast of the proposed method.}

\item{u}{The quantile.
Four high quantiles are considered, i.e. 0.750, 0.835, 0.975, and 0.995.}
}
\value{
The average SPL error of all time series.
}
\description{
Calculate the SPL error of the proposed method
}
